requirements.txt
asgiref==3.11.0
attrs==25.4.0
celery==5.5.3
colorama==0.4.6
coverage==7.12.0
Django==5.2.8
django-cors-headers==4.9.0
django-filter==24.3
django-jazzmin==3.0.1
django-redis==6.0.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
drf-spectacular==0.29.0
factory_boy==3.3.3
Faker==38.2.0
inflection==0.5.1
iniconfig==2.3.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
packaging==25.0
pillow==12.0.0
pluggy==1.6.0
psycopg2-binary==2.9.11
Pygments==2.19.2
pytest==9.0.1
pytest-cov==7.0.0
pytest-django==4.11.1
python-decouple==3.8
PyYAML==6.0.3
redis==7.1.0
referencing==0.37.0
rpds-py==0.30.0
sqlparse==0.5.4
typing_extensions==4.15.0
tzdata==2025.2
uritemplate==4.2.0

# websocket - Django Channels
channels==4.3.1
channels_redis==4.3.0
daphne==4.2.1
whitenoise==6.11.0
